Implement q_delete_dup with two pointers
Use two pointers to identify the range of nodes with
same value, and release their memory afterward.

We don't release the memory on the way because we won't be
able to tell if duplicate values exist.

NOTE: One shouldn't use list_for_each to iterate through a
list if you are going to modify it, since it will cause undefined
behavior.

Change-Id: I31a0fdee043b6c9ab7ec734c9f7ed242cd32139d