Implementing q_reverseK by counting first.
reverseK is done by the same procedure as reverse the whole list,
which is swap the 'prev' and 'next' pointers on each node. However,
this time, we have to make sure there are enough nodes to make a
group of K, also we need to take care of the head of the previous group
and the tail of the current group.

Change-Id: I0e7c1a875297f3c427ced866a3e046afc5913c8e