Implement mergesort by taking out the dummy head
Remove the dummy head of the list, and covert the circular
doubly linked list into a doubly linked list by unlinking the head
node.

Create a helper function to pass the new list into the
q_mergesort() so that we don't need to take of the dummy heads.

Divide the list into sublists until only one element remains.
and use q_mergelists() to combine the sorted sublists.

To sort the list in descending order, I sort it in ascending order
and simply reverse it. It's probably not a good approach but gets
thing done for now.

Change-Id: I573eb6722708e053c795ea0d18cf8ef44ce59f89