Implementing q_ascend and q_descend
The function  iterates from the tail,
keeping track of the smallest encountered value,
and deletes nodes that have a greater value.

After the iteration, we can make sure the remaining list
is in ascending order because every node won't be smaller than
the nodes to the right.

We have the same approach for q_descend but this time,
we iterate from the tail, and delete nodes that have a smaller value.

Change-Id: Ia50a846530cb31135da0022baf7fccc42690b1ff