Implemeting q_ascend and q_descend
For q_ascend, iterate through the list from the tail to remove any
node that is bigger than the minimum value, and keep track of the
minimum value during the iteration.

After the iteration, we can make sure that every node has no nodes
with smaller value to the right side of it.

For q_descend, we use the same approach, but to remove any node
that is smaller than the maximum value.

Change-Id: I2072b4c1279291b7f116c63550549ed5a568f955