Remove redundant explicit casting in q_new
As shown in C99 [6.3.2.3] and C99 [7.20.3]

"A pointer to void may be converted to or from a pointer to any
incomplete or object type. A pointer to any incomplete or object
type may be converted to a pointer to void and back again."

"The pointer returned if the allocation succeeds is suitably
aligned so that it may be assigned to a pointer to any type
of object and then used to access such an object or an array
of such objects in the space allocated."

'malloc' will return a pointer to void and it requires no explicit
casting.

I've removed the casting process and redundant empty lines as well.

Change-Id: I9f5fa46206e6d76587f7a49a83759b398248615a