Remove delete operation in q_free
The purpose of q_free is to free all storage used by queue,
rather than delete a single node, so there's no need to maintain
the whole linked list.

I was wondering whether I should release the memory used by the linked
list inside the structure, and I found that it is part of the structure
, and its life cycle will end when the "element_t" is freed. However,
any member that is a pointer should be released using free.

It is not necessary to initialize "entry" and "safe" to NULL, since
they will be assigned with values in the loop.