Use strdup to duplicate a string
'strdup' performs the same function as my previous implementation,
which is declare a pointer to char and use 'memcpy' to make a copy,
so I replace it to make the code more concise.

strdup returns a pointer to the duplicated string, and if
memory allocation fails, it returns NULL.

Added additional checks to ensure all memory allocation succeed.

Change-Id: I6ffe0a906eebc6b0451287df7ea6fac39a60b94c