Reuse q_insert_head in q_insert_tail
Our goal is to insert the node between head and head->prev, we can
achieve it by reusing q_insert_head function and passing head->prev
as the argument.

This approach works even when the queue contains only the head node
because if the first inserted node will become both head->next and
head->prev.

Change-Id: I371404fd3681aeffc20edff1e08aabe14b3ded4b