Check buffer size before copying string
In q_remove_head, we should copy the removed string to '*sp'. However,
we have to ensure that the size of the string does not exceed the buffer
size.

This time, we use strncpy to copy exact buffer - 1 characters from
ele->value to '*sp' and add a null terminator at the end.

Change-Id: I9c0b88579dfdd2ce099f03efd3d928072e6a028b