Reuse q_remove_head in q_remove_tail
It's obvious that, for any queue containing n >= 2 nodes, we can pass
'head->prev-prev' into the function to remove the tail element of the
queue.

For queue of size one, 'head->prev->prev' points to head, and in this
case, removing the head element is the same thing as removing the tail
one.

Change-Id: I8700cc8b11fd90611b2f9826dd36964cfa960d3e