Fix logic and memory release in q_ascend
We don't have to check for the minimum value every time we encounter a
new node. If the left node not greater than the current node, it must be
the new minimum value.

I release the memory of the list_head structure, and the
correct way to release it is by releasing the whole element_t
structure.

Change-Id: I0678e5a211c96406ddf9e05c554d3c634a178d6b