Rewrite iterative merge sort, and q_merge_lists
I implemented merge sort recursively. However, it didn't pass
trace-test-14 because it has too much function call and takes up too
much space.

I did not come up with this approach, and it's a copy of @rota1001's
code.

We use an stack to mimic function calls and this time, we don't split
the list, we insert each node into the stack instead. If the top two
group of nodes has the same size, we can combine them by their value
and put them back into the stack. After inserting all the nodes, we
combine all the remaining groups in the stack.

Before inserting any node into the stack, we should unlink their next
pointer so that it's easier to process in q_merge_lists.

After all the operation, we should link all the nodes together with
'list_add_tail' and turn it back to a doubly circular linked list.

We make sure each group is sorted by doing steps above, and
it's the fundamental idea of merge sort. In fact, I found this approach
really smart because it's more intuitive.

Change-Id: If9ae68a8155a285c477869286197720c095261bf